requirements.txt
explainerdashboard==0.4.1.1
keras==2.11.0
matplotlib==3.6.3
missingno==0.5.1
numpy==1.23.5
pandas==1.5.3
pandas_profiling==3.6.6
Pillow==9.4.0
plotly==5.13.0
scikit_learn==1.2.1
seaborn==0.12.2
streamlit==1.17.0
streamlit_pandas_profiling==0.1.3
ydata_profiling==4.0.0
